requirements.txt
requests
pandas
numpy
matplotlib
scikit-learn
nltk
textblob
python-dotenv
